feat: New config system
