+++ README.md
@@ -18,7 +18,7 @@ import {NgFluxifyModule} from 'ng-fluxify;
 @NgModule({
   imports: [
     ...
-    NgFluxifyModule.initialize({enableStoreLogger: true}),
+    NgFluxifyModule,
     ...
   ],
   ...
