+++ ng-package.prod.json
@@ -4,7 +4,8 @@
   "whitelistedNonPeerDependencies": [
     "angular",
     "redux",
-    "rxjs"
+    "rxjs",
+    "immutable"
   ],
   "lib": {
     "entryFile": "src/public_api.ts"
