+++ package.json
@@ -26,7 +26,7 @@
     "@angular-redux/form": "^9.0.1",
     "@angular-redux/store": "^9.0.0",
     "rxjs": "~6.4.0",
-    "immutable": "^3.8.0"
+    "immutable": "~3.8.0"
   },
   "dependencies": {
     "@angular-redux/form": "^9.0.1",
@@ -35,7 +35,7 @@
     "redux-logger": "^3.0.6",
     "redux-observable": "^1.0.0",
     "rxjs": "~6.4.0",
-    "immutable": "^3.8.0"
+    "immutable": "~3.8.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^7.0.0",
