+++ src/lib/ng-fluxify.module.ts
@@ -1,12 +1,9 @@
-import {InjectionToken, Injector, ModuleWithProviders, NgModule} from '@angular/core';
+import {Injector, ModuleWithProviders, NgModule} from '@angular/core';
 import {NgRedux, NgReduxModule} from '@angular-redux/store';
-import {IAppState, RootReducer} from './stores';
 import {HttpClientModule} from '@angular/common/http';
 import {EntityDescriptor} from './domain/descriptors';
-import {logger} from 'redux-logger';
-import {applyMiddleware} from 'redux';
-
-const NgFluxifyConfigService = new InjectionToken<NgFluxifyConfig>('NgFluxifyConfigService');
+import {IAppState} from './stores';
+import {NgFluxifyConfig, NgFluxifyConfigService, NgReduxService} from './services';
 
 // @dynamic
 @NgModule({
@@ -15,36 +12,18 @@ const NgFluxifyConfigService = new InjectionToken<NgFluxifyConfig>('NgFluxifyCon
     NgReduxModule
   ],
   declarations: [],
+  providers: [NgReduxService],
   exports: []
 })
 export class NgFluxifyModule {
   static injector: Injector;
   static ngRedux: NgRedux<IAppState>;
 
-  static get ngFluxifyConfigService(): Promise<NgFluxifyConfig> {
-    return new Promise<NgFluxifyConfig>(resolve => {
-      const interval = setInterval(() => {
-        console.log(this.injector.get(NgFluxifyConfigService, false));
-        if (this.injector.get(NgFluxifyConfigService, false)) {
-          if (interval) {
-            clearInterval(interval);
-          }
-          resolve(this.injector.get(NgFluxifyConfigService, {}));
-        }
-      }, 5);
-    });
-  }
-
   private static entityList: Map<string, EntityDescriptor> = new Map<string, EntityDescriptor>();
-  private static isRootStoreConfigured: boolean;
 
   constructor(public ngRedux: NgRedux<IAppState>, private injector: Injector) {
     NgFluxifyModule.injector = injector;
     NgFluxifyModule.ngRedux = ngRedux;
-
-    if (NgFluxifyModule.entities.length) {
-      NgFluxifyModule.configureStore();
-    }
   }
 
   public static initialize(ngFluxifyConfig: NgFluxifyConfig): ModuleWithProviders<NgFluxifyModule> {
@@ -63,38 +42,15 @@ export class NgFluxifyModule {
     return Array.from(NgFluxifyModule.entityList.values());
   }
 
-  public get entitites(): EntityDescriptor[] {
+  public get entities(): EntityDescriptor[] {
     return NgFluxifyModule.entities;
-  }
-
-  private static configureStore() {
-    this.ngFluxifyConfigService.then(ngFluxifyConfig => {
-      const enhancers = [];
-
-      if (ngFluxifyConfig.enableStoreLogger) {
-        enhancers.push(applyMiddleware(logger));
-      }
-
-      this.ngRedux.configureStore(RootReducer.getReducer(NgFluxifyModule.entities), {}, [], []);
-    });
-
-    this.ngRedux.configureStore(RootReducer.getReducer(NgFluxifyModule.entities), {}, [], []);
-    this.isRootStoreConfigured = true;
   }
 
   public static registerEntity(entityDescriptor: EntityDescriptor) {
     NgFluxifyModule.entityList.set(entityDescriptor.name, entityDescriptor);
 
-    if (NgFluxifyModule.ngRedux) {
-      if (this.isRootStoreConfigured) {
-        this.ngRedux.replaceReducer(RootReducer.addReducer(entityDescriptor));
-      } else {
-        this.configureStore();
-      }
+    if (this.injector.get<NgReduxService>(NgReduxService, null)) {
+      this.injector.get<NgReduxService>(NgReduxService).registerEntity(entityDescriptor);
     }
   }
-}
-
-export interface NgFluxifyConfig {
-  enableStoreLogger?: boolean;
 }
