+++ src/lib/services/index.ts
@@ -1,2 +1,4 @@
 export * from './IEntity.service';
+export * from './ng-fluxify-config.service';
+export * from './ng-redux.service';
 export * from './rest-service';
