+++ src/lib/services/ng-fluxify-config.service.ts
@@ -1,0 +1,9 @@
+import {InjectionToken} from '@angular/core';
+
+export const NgFluxifyConfigService = new InjectionToken<NgFluxifyConfig>('NgFluxifyConfig', {
+  factory: () => <NgFluxifyConfig>{}
+});
+
+export interface NgFluxifyConfig {
+  enableStoreLogger?: boolean;
+}
